fix: parseDouble nonisolated, iOS refresh reloads UI, OpenAI usage balance, unify Int64
- ApiPoller.parseDouble: add nonisolated (Swift 6 concurrency)
- iOS CloudDataService.refresh(): track currentRange, call loadSnapshot
  after fetch so pull-to-refresh actually updates the UI
- OpenAI balance: for .usage providers, store cumulative spend as
  negative in balance_snapshot so balanceDailySpend detects increases
  as spend. Fix coin sound detection for usage type.
- DashboardSnapshot: unify todayCalls/todayTokens as Int64 (macOS
  was Int, shared was Int64 — now both Int64)

+++ Sources/Store/StatsService.swift
@@ -615,8 +615,8 @@ enum StatsService {
         let dailyPts = st.map { TrendPoint(ts: $0.date.timeIntervalSince1970, value: $0.cost, calls: $0.calls, tokens: $0.tokens, netLines: $0.netLines) }
         let codePts = cd.map { TrendPoint(ts: $0.date.timeIntervalSince1970, value: Double($0.added), calls: 0, tokens: 0, netLines: $0.added - $0.deleted, added: $0.added, deleted: $0.deleted) }
         let balPts = Dictionary(grouping: bl, by: { $0.date }).compactMap { d, v in TrendPoint(ts: d.timeIntervalSince1970, value: v.reduce(0) { $0 + $1.spend }, calls: 0, tokens: 0, netLines: 0) }
-        let todayCall = st.reduce(0) { $0 + $1.calls }
-        let todayTok = st.reduce(0) { $0 + $1.tokens }
+        let todayCall = Int64(st.reduce(0) { $0 + $1.calls })
+        let todayTok = Int64(st.reduce(0) { $0 + $1.tokens })
 
         return DashboardSnapshot(
             todayCost: tc, weekCost: wc, monthCost: mc,
